refactor: common code from ChannelInfo/Description -> BaseInfoFragment
